Don't panic or log fatal in the library code (#16)
In order to accommodate forward and backward compatibility in the parser, parsing errors cannot be fatal or interrupt the main program. Instead, throw them in an error channel and let the client code decide what to do. Many parsing errors, like missing or extra fields, are not going to be reason to stop listening.