Add send integration test
Debugging this brought up an opportunity for better exception behavior. When trying to send a command, if we haven't heard from WSJTX yet, we don't know where to send the command. Now the server returns a better error message to the client code and exposes it in the API for checking with `errors.Is()`.